+++ kataphron/weapons/__init__.py
@@ -1,0 +1,3 @@
+from .base import *
+from .melee import *
+from .ranged import *
