+++ kataphron/weapons/ranged.py
@@ -1,45 +1,13 @@
-import abc
-from dataclasses import dataclass
-from typing import Sequence
+from kataphron.weapons.base import RangeWeapon, Ability
 
-
-@dataclass
-class Ability:
-    name: str
-    description: str | None = None
-    # * this can now do much more than a simple string
-    times_used = 0
-
-class Weapon:
-    name: str
-    range: int
-    type: str
-    S: int | str  # FIXME: this should be int or str or custom type
-    AP: int
-    D: int | str  # FIXME: this should be int or str or custom type
-    abilities: Sequence[Ability] = []
-
-    def __str__(self):
-        return f'{self.name} | {self.range} | {self.type} | {self.S} | {self.AP} | {self.D} | {self.abilities}'
-    
-    @abc.abstractmethod
-    def use(self, with_ability: Ability|None = None) -> None:
-        raise NotImplementedError()
-
-
-class RangeWeapon(Weapon):
-
-    def __init__(self, abilities: Sequence[Ability] | Ability):
-        if isinstance(abilities, Ability):
-            abilities = [abilities]
-        self.abilities = abilities
-
-    def use(self, with_ability: Ability|None = None): # * I wouldn't say this is perfectly valid way to do that, that's just an example
-        if with_ability is None:
-            return
-        if with_ability not in self.abilities:
-            raise AttributeError("Invalid Ability.")
-        with_ability.times_used += 1
+__all__ = ["CognisFlamer",
+           "HeavyGravCanon",
+           "PlasmaCulverin",
+           "PlasmaCulverinSupercharged",
+           "PhosphorBlaster",
+           "HeavyArcRifle",
+           "TorsionCannon",
+           ]
 
 
 class CognisFlamer(RangeWeapon):
@@ -125,33 +93,3 @@ class TorsionCannon(RangeWeapon):
         self.S = 8
         self.AP = -4
         self.D = "D3+3"
-
-
-class MeleeWeapon(Weapon):
-    def __init__(self, abilities: Ability | Sequence[Ability]):
-        if isinstance(abilities, Ability):
-            abilities = [abilities]
-        self.abilities = abilities
-        self.range = 0
-        self.type = "Melee"
-
-
-class ArcClaw(MeleeWeapon):
-    def __init__(self):
-        super().__init__(
-            Ability("No name", "Each time the bearer fights, it makes 1 additional attack with this weapon."
-                    "Each time an attack is made with this weapon against a VEHICLE unit, that attack has a Damage"
-                    "characteristic of 2 and an unmodified wound roll of 4+ successfully wounds the target."))
-        self.name = "Arc claw"
-        self.S = "+1"
-        self.AP = -3
-        self.D = 1
-
-
-class HydraulicClaw(MeleeWeapon):
-    def __init__(self):
-        super().__init__([])
-        self.name = "Hydraulic claw"
-        self.S = "x2"
-        self.AP = -2
-        self.D = 3
